First fix

+++ build_forest.m
@@ -1,6 +1,6 @@
 function [forest] = build_forest( data_labels, num_trees )
     forest = {};
-    dimensions = 5;
+    dimensions = uint32(5);
     for i=1:num_trees
         indices = randsample(size(data_labels, 1), size(data_labels, 1), true);
         data = data_labels(indices, :);
